+++ src/main/java/kr/hhplus/be/server/application/dataplatform/DataPlatformClient.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.application.dataplatform;
 
-import kr.hhplus.be.server.application.event.OrderEvent;
+import kr.hhplus.be.server.domain.order.event.OrderEvent;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class DataPlatformClient {
 
-    public void sendOrderInvoice(OrderEvent.Complete event) {
+    public void sendOrderInvoice(OrderEvent.Completed event) {
         try {
             log.info("[OrderEvent] 외부 플랫폼 전용 MockAPI 주문 정보 전송- orderId: {}", event.order().orderId());
