+++ src/main/java/kr/hhplus/be/server/application/order/listener/OrderCompletedEventListener.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.server.domain.order.event;
+package kr.hhplus.be.server.application.order.listener;
 
 import kr.hhplus.be.server.application.dataplatform.DataPlatformClient;
-import kr.hhplus.be.server.application.event.OrderEvent;
+import kr.hhplus.be.server.domain.order.event.OrderEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -12,13 +12,13 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @Slf4j
 @RequiredArgsConstructor
 @Component
-public class OrderCompleteEventListener {
+public class OrderCompletedEventListener {
 
     private final DataPlatformClient dataPlatformClient;
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     @Async
-    public void handle(OrderEvent.Complete event) {
+    public void handle(OrderEvent.Completed event) {
         dataPlatformClient.sendOrderInvoice(event);
     }
 }
