+++ src/main/java/kr/hhplus/be/server/domain/order/event/OrderEvent.java
@@ -1,10 +1,10 @@
-package kr.hhplus.be.server.application.event;
+package kr.hhplus.be.server.domain.order.event;
 
 import kr.hhplus.be.server.domain.order.Order;
 
 public class OrderEvent {
 
-    public record Complete(Order order) {}
+    public record Completed(Order order) {}
 
     public record Canceled(Order order) {}
